I have no clue how much i changed in this.
But i remember createing an exercide home page :)
and fixing a bug that deleted all my data :D

+++ app/src/main/java/io/github/michael_bailey/gym_log_book/data_type/ExerciseItem.kt
@@ -5,10 +5,10 @@ import java.time.LocalDate
 
 /// This represents an item in the exercise list
 data class ExerciseItem(
-	override val id: Long,
+	override val id: Int,
 	val date: LocalDate,
 	var exercise: String,
 	var setNumber: Int,
-	var weight: Int,
+	var weight: Double,
 	var reps: Int,
 ) : Identifiable()
